Refactor project add build package
