+++ README.md
@@ -1,0 +1,19 @@
+# Minesweeper
+
+
+## Installation
+
+```
+$ pip install minesweeper-py
+```
+
+## Start game
+```
+$ minesweeper
+```
+
+## License
+Minesweeper is released under the MIT license. See LICENSE for details.
+
+## Contact
+Follow me on GitHub [@Gerrux](https://github.com/Gerrux)
