+++ build.sh
@@ -1,0 +1,3 @@
+python setup.py bdist_wheel
+python setup.py sdist
+rm -rf build
