+++ minesweeper/assets/colors.py
