+++ minesweeper/assets/sprites/blank-block.png
