+++ minesweeper/assets/sprites/flag.png
