+++ minesweeper/assets/sprites/warn2.png
