+++ minesweeper/assets/sprites/warn3.png
