+++ minesweeper/assets/sprites/warn4.png
