+++ minesweeper/assets/sprites/warn5.png
