+++ minesweeper/assets/sprites/warn6.png
