+++ minesweeper/assets/sprites/warn7.png
