+++ minesweeper/assets/sprites/warn8.png
