+++ minesweeper/controller/controller.py
@@ -10,7 +10,7 @@ class MinesweeperController:
     def get_view_state(self):
         return self.view_state
 
-    def switch_screen(self, state):
+    def set_view_state(self, state):
         self.view_state = state
 
     def set_game_mode(self, mode):
@@ -27,7 +27,9 @@ class MinesweeperController:
         try:
             self.model.start_game(*map(int, game_settings))
         except:
-            self.model.start_game(self.model.row_count, self.model.column_count, self.model.bomb_count)
+            self.model.start_game(
+                self.model.row_count, self.model.column_count, self.model.bomb_count
+            )
 
         self.view.create_board()
 
@@ -41,5 +43,7 @@ class MinesweeperController:
 
     def on_right_click(self, row, column):
         self.model.next_cell_mark(row, column)
-        self.view.block_cell(row, column, self.model.get_cell(row, column).state == 'flagged')
+        self.view.block_cell(
+            row, column, self.model.get_cell(row, column).state == "flagged"
+        )
         self.view.sync_with_model()
