+++ minesweeper/model/__init__.py
@@ -1,0 +1,3 @@
+from .game_model import MinesweeperModel
+
+__all__ = ["model"]
