+++ minesweeper/model/cell_model.py
@@ -1,4 +1,3 @@
-
 class MinesweeperCell:
     # Возможные состояния игровой клетки:
     #   closed - закрыта
@@ -9,11 +8,11 @@ class MinesweeperCell:
     def __init__(self, row, column):
         self.row = row
         self.column = column
-        self.state = 'closed'
+        self.state = "closed"
         self.mined = False
         self.counter = 0
 
-    mark_sequence = ['closed', 'flagged', 'questioned']
+    mark_sequence = ["closed", "flagged", "questioned"]
 
     def next_mark(self):
         if self.state in self.mark_sequence:
@@ -21,5 +20,5 @@ class MinesweeperCell:
             self.state = self.mark_sequence[(state_index + 1) % len(self.mark_sequence)]
 
     def open(self):
-        if self.state != 'flagged':
-            self.state = 'opened'
+        if self.state != "flagged":
+            self.state = "opened"
