+++ minesweeper/view/elements/image.py
