+++ minesweeper/view/elements/rectangle.py
