+++ minesweeper/view/main_view.py
@@ -1,5 +1,5 @@
-from views.game_view import MinesweeperView
-from views.menu_view import MenuView
+from .views.game_view import MinesweeperView
+from .views.menu_view import MenuView
 
 
 class MainView:
@@ -13,11 +13,15 @@ class MainView:
         self.current_view.run()
 
     def switch_on_game(self):
-        self.current_view = MinesweeperView(self.model, self.controller, self.controller.get_game_mode())
+        self.current_view = MinesweeperView(
+            self.model, self.controller, self.controller.get_game_mode()
+        )
         self.current_view.run()
 
     def run(self):
         while True:
             match self.controller.get_view_state():
-                case "menu": self.switch_on_menu()
-                case "game": self.switch_on_game()
+                case "menu":
+                    self.switch_on_menu()
+                case "game":
+                    self.switch_on_game()
